# custom/laravel/docs/openapi/paths/super_admin.yaml
openapi: 3.0.0
info:
  title: Super Admin API
  description: API for Super Admin operations with 100% Rails parity
  version: 2.0.0
servers:
  - url: 'https://api.example.com/v1'
paths:
...
  /super_admin/settings:
    get:
      tags:
        - Super Admin Settings
      summary: List all installation settings
      operationId: listSuperAdminSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of settings
    patch:
      tags:
        - Super Admin Settings
      summary: Update multiple settings
      operationId: updateSuperAdminSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - settings
              properties:
                settings:
                  type: object
                  additionalProperties: true
                force:
                  type: boolean
      responses:
        '200':
          description: Settings updated
        '207':
          description: Partial success with errors
    post:
      tags:
        - Super Admin Settings
      summary: Create new setting
      operationId: createSuperAdminSetting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - value
              properties:
                name:
                  type: string
                value:
                  oneOf:
                    - type: string
                    - type: number
                    - type: boolean
                    - type: object
                locked:
                  type: boolean
      responses:
        '201':
          description: Setting created
...
  /super_admin/account_users:
    get:
      tags:
        - Super Admin Account Users
      summary: List account-user relationships
      operationId: listSuperAdminAccountUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: account_id
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
            enum: [agent, admin]
        - name: availability
          in: query
          schema:
            type: integer
            enum: [0, 1]
        - name: search
          in: query
          schema:
            type: string
        - name: account_search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of account-user relationships
    post:
      tags:
        - Super Admin Account Users
      summary: Create account-user relationship
      operationId: createSuperAdminAccountUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - account_id
                - role
              properties:
                user_id:
                  type: integer
                account_id:
                  type: integer
                role:
                  type: string
                  enum: [agent, admin, "1", "2"]
                availability:
                  type: integer
                  enum: [0, 1]
                settings:
                  type: object
      responses:
        '201':
          description: Account-user relationship created
        '409':
          description: User already associated with account
...